requirements.txt
aioredis==1.3.1
aliyun-python-sdk-alidns==2.0.7
aliyun-python-sdk-core-v3==2.9.0
amqp==2.5.1
ansible==2.8.5
ansible-stubs==0.1.dev1
asgiref==3.2.7
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
backports.csv==1.0.7
bcrypt==3.1.7
billiard==3.6.1.0
blessings==1.7
bpython==0.18
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
curtsies==0.3.0
daphne==2.5.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-extensions==3.0.8
django-filter==2.2.0
django-filter-stubs==0.1.1
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.4
django-widget-tweaks==1.4.5
djangorestframework==3.10.2
djangorestframework-stubs==1.1.0
dnspython==1.16.0
docker==4.0.2
dwebsocket==0.5.12
et-xmlfile==1.0.1
future==0.17.1
gitdb2==2.0.6
GitPython==3.0.3
google-auth==1.6.3
greenlet==0.4.15
harbor-py-light==2.1.2
hiredis==1.0.1
httpie==1.0.2
httplib2==0.13.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.23
incremental==17.5.0
IPy==1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.9.4
kombu==4.6.5
kubernetes==10.0.1
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==0.6.2
multi-key-dict==2.0.3
mysqlclient==1.4.2.post1
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.6.2
paramiko==2.4.2
pbr==5.4.4
Pillow==6.1.0
prettytable==0.7.2
psutil==5.6.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.9.0
pydot==1.4.1
pygame==1.9.6
Pygments==2.4.2
PyHamcrest==2.0.2
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypinyin==0.35.4
python-dateutil==2.8.0
python-gitlab==1.12.1
python-jenkins==1.7.0
pytz==2019.1
PyYAML==5.1.1
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
service-identity==18.1.0
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
tablib==0.13.0
Twisted==20.3.0
txaio==20.4.1
typing==3.7.4
urllib3==1.25.3
uWSGI==2.0.17
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==0.6.0
zope.interface==5.1.0
